Add Lebanon.  Also, more refactoring: add vote modes to Globals, move createRecord to TemplateRecord.
