+++ ElectionGlobals.py
@@ -8,3 +8,9 @@ class Globals:
     RESULTS_DIR = 'Results_PDF'
     # Per-count list of URLs corresponding to this county
     URL_LIST_FILENAME = "URL_List.txt"
+
+    # Modes of vote result
+    MODE_ELECTION_DAY = 'ELECTION DAY'
+    MODE_MAIL_IN = 'MAIL-IN'
+    MODE_PROVISIONAL = 'PROVISIONAL'
+    MODE_TOTAL = 'TOTAL'
