+++ PA/Lebanon/Results_PDF/URL_List.txt
@@ -1,0 +1,1 @@
+https://www.lebanoncountypa.gov/getmedia/24db601e-d2fc-4e2c-863c-e55951ae51ed/2024-Presidential-Precinct-Summary-Report_Final.pdf
