+++ README.md
@@ -1,12 +1,14 @@
 # elections
 Scripts for parsing PDF election results for the 2024 Election.
 
-Developed with Python 3.
+Developed using Python 3.
 
 Requires installing 'pypdf' package: `pip3 install pypdf`
 
-## How to run
+## How to generate CSV results:
 ```
 python3 readFulton.py
 python3 readHuntingdon.py
+python3 readJuniata.py
+python3 readLebanon.py
 ```
