+++ readLebanon.py
@@ -1,0 +1,4 @@
+from readSingleDoc import readCountyResults
+from ElectionGlobals import Globals
+
+readCountyResults(Globals.STATE, Globals.STATE_ABBREV, 'Lebanon')
